+++ packages/expo-modules-core/CHANGELOG.md
@@ -15,6 +15,7 @@
 ### 💡 Others
 
 - Update edge-to-edge package to call `updateEdgeToEdgeFeatureFlag` ([#46335](https://github.com/expo/expo/pull/46335) by [@zoontek](https://github.com/zoontek))
+- `NativeArrayBuffer` arguments no longer copy the buffer when it's already native-backed. ([#46448](https://github.com/expo/expo/pull/46448) by [@barthap](https://github.com/barthap))
 
 ## 56.0.13 — 2026-05-26